# perf(compose): optimize `await`