Do not `await` if the handler is not Promise in `compose`.

## Benchmark Results

Although all routes are "sync", I could see a large improvement in performance.

```
hono - trie-router(default) x 393,919 ops/sec ±4.52% (86 runs sampled)
hono - regexp-router x 478,140 ops/sec ±2.85% (83 runs sampled)
hono optimized - trie-router(default) x 427,940 ops/sec ±4.32% (81 runs sampled)
hono optimized - regexp-router x 512,488 ops/sec ±5.28% (75 runs sampled)
Fastest is hono optimized - regexp-router
✨  Done in 28.36s.
```

## Deno

Not so different on Deno.

```
Before: Requests/sec: 50226.7374
After: Requests/sec: 50296.9231
```

## Bun

On Bun, will be 15% faster.

```
Before: Requests/sec: 57043.0018
After:  Requests/sec: 65658.9860
```

## Note

The test on the bottom in `hono.test.ts` was wrong. So, I've rewritten, and all tests are passed.